@@ -12,24 +12,26 @@
     "test:watch": "jest --watch",
     "prettify": "prettier --write ./src/**/*.{ts,js}"
   },
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
     "@types/jest": "^24.0.0",
     "@types/node": "^10.12.21",
-    "eslint": "^5.13.0",
+    "@typescript-eslint/eslint-plugin": "^1.3.0",
+    "@typescript-eslint/parser": "^1.3.0",
+    "eslint": "^5.14.0",
     "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-airbnb-base": "^13.1.0",
     "eslint-config-prettier": "^4.0.0",
     "eslint-import-resolver-typescript": "^1.1.1",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.4",
     "jest": "^24.1.0",
     "nodemon": "^1.18.9",
     "prettier": "^1.16.4",
     "ts-jest": "^23.10.5",
     "ts-node": "^8.0.2",
-    "tslint": "^5.12.1",
     "typescript": "^3.3.1"
   }
 }
