@@ -1,4 +1,4 @@
-const { add } = require('./example');
+import { add } from './example';
 
 test('basic test', () => {
   expect(add(2, 4)).toBe(6);
