@@ -3,5 +3,6 @@ function add(x, y) {
 }
 
 module.exports = {
-  add
+  add,
+  jsadjsahj
 };
